Tap to build UI and Railgun and chainBeam
Tap to build UI:
Tapping on the center opens up buttons

Railgunbeam:
ModifyBooleans to have the beam shoot. You can use it on the turret itself or Instantiate it as a projectile

ChainBeam:
LineRenderer chains to nearest few objects. Ronnie is a sample texture.

my codes do not have comments because i suck